Code review comment changes, segregating S3 and Cleversafe code.
