+++ pom.xml
@@ -13,7 +13,7 @@
 
 	<groupId>eu.elixir.ega.ebi</groupId>
 	<artifactId>ega-data-api</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
+	<version>1.0.1-SNAPSHOT</version>
 	<packaging>pom</packaging>
 	<name>EGA Data API</name>
 	<description>EGA Data API With Multiple Maven Modules</description>
